+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Offline playback support on Android and iOS
 - `SourceConfig.options` to enable configuring stream start position
 - `PlayerConfig.adaptationConfig` to allow configuring the player's adaptation behavior
+- `Player.setMaxSelectableBitrate` to allow setting the maximum selectable bitrate on the `Player`
 
 ## [0.9.2] (2023-08-24)